fix(renamed packcage): #84
