+++ src/main/java/acme/features/any/course/AnyCourseController.java
@@ -1,5 +1,5 @@
 
-package acme.features.any.courses;
+package acme.features.any.course;
 
 import javax.annotation.PostConstruct;
